+++ test/test_choice_implicit_delete.sh
@@ -78,7 +78,7 @@ CLICON_PLUGIN="example_cli";
 # Autocli syntax tree operations
 set @datamodel, cli_auto_set();
 delete("Delete a configuration item") {
-      @datamodel, @add:leafref-referred, cli_auto_del();
+      @datamodel, @add:leafref-no-refer, cli_auto_del();
       all("Delete whole candidate configuration"), delete_all("candidate");
 }
 validate("Validate changes"), cli_validate();